# src/docs/swagger.yaml
basePath: /v1
definitions:
...
  main.UserResponse:
    properties:
      created_at:
        type: string
      dailing_code:
        type: string
        x-nullable: "true"
      email:
        type: string
        x-nullable: "true"
      is_onboarding_completed:
        type: boolean
      last_name:
        type: string
        x-nullable: "true"
      name:
        type: string
        x-nullable: "true"
      phone:
        type: string
        x-nullable: "true"
      second_last_name:
        type: string
        x-nullable: "true"
      updated_at:
        type: string
      user_id:
        type: string
      username:
        type: string
        x-nullable: "true"
      verified_email:
        type: boolean
      verified_phone:
        type: boolean
    type: object
...
  utils.Response-any:
    properties:
      data: {}
      exception:
        type: string
        x-nullable: "true"
      message:
        type: string
      status_code:
        type: integer
    type: object
...
  utils.Response-main_UserResponse:
    properties:
      data:
        $ref: '#/definitions/main.UserResponse'
      exception:
        type: string
        x-nullable: "true"
      message:
        type: string
      status_code:
        type: integer
    type: object
...
host: api.smartgo.com
info:
  contact: {}
  description: SmartGo API
  title: SmartGo API
  version: "1.0"
paths:
...
  /users/{id-user}:
    get:
      consumes:
      - application/json
      description: This endpoint gets the user profile for a user.
      parameters:
      - description: User ID formatted as UUID
        in: path
        name: id-user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-main_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - users
  /users/{id-user}/onboarding:
    patch:
      consumes:
      - application/json
      description: This endpoint completes the onboarding process for a user.
      parameters:
      - description: User ID formatted as UUID
        in: path
        name: id-user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      security:
      - BearerAuth: []
      summary: Complete onboarding
      tags:
      - users
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"